feat: Allow chunk to swap previous chunk
By default, when a new chunk arrives, it will be appended right after
the previous chunk. This commit introduces a new element tag
hx-chunked-mode that controls how the chunks are inserted. By default it
will be append as before but if set to 'swap' it will swap out the
previous chunk.